Capstone Project: Data reading script
Check in the script to read the CSV files and compile into a large data
frame for each year. The SMART data is ignored at this point due to the
large size.
There are a few limitations, i.e. the data frame is not saved with the
name of the corresponding year.